+++ libs/opengl/shaders/points.f.glsl
@@ -1,12 +1,12 @@
 R"XXX(
-#version 330 core
+#version 300 es
 
 // FRAGMENT SHADER: Default shader for MRPT CRenderizable objects
 // Jose Luis Blanco Claraco (C) 2019-2020
 // Part of the MRPT project
 
-out vec4 color;
-in vec4 frag_color;
+out highp vec4 color;
+in highp vec4 frag_color;
 
 void main()
 {